FS-4674: add `funding_type` attribute to fund model
Adds the `FundingType` enum to the fund table as described in the ticket.
For the time being allow the `funding_type` column to be nullable
to allow appropriately backfilling the field in existing environments.

We may want to decide to not allow this field to have a null value for
creating new funds, which can be a subsequent migration if required.