FS-4674: alembic migration for funding type model
Autogenerates the alembic migration needed for the `funding_type` column.

There was an issue running the upgrade with alembic having not autogenerated
the postgres "type" for the enum. According to the linked github issue
this is known for updates. It's been resolved here by explicitly adding
prompts to create and drop the type on upgrade and downgrade.